+++ scripts/session_saver.sh
@@ -22,8 +22,10 @@ dump() {
 }
 
 save_all_sessions() {
-	mkdir -p $HOME/.tmux
-	dump > $HOME/.tmux/session
+	local session_path="$(session_path)"
+	mkdir -p "$SESSIONS_DIR"
+	dump > $session_path
+	ln -fs "$session_path" "$(last_session_path)"
 	display_message "Saved all Tmux sessions!"
 }