+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- saving a new session does not remove the previous one
 
 ### v0.0.1, 2014-08-26
 - started a project