+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### master
 - saving a new session does not remove the previous one
 - make the directory where sessions are stored configurable
+- support only Tmux v1.9 or greater
 
 ### v0.0.1, 2014-08-26
 - started a project