+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - make the directory where sessions are stored configurable
 - support only Tmux v1.9 or greater
 - display a nice error message if saved session file does not exist
+- added README
 
 ### v0.0.1, 2014-08-26
 - started a project